refactor(sync): split publisher and consumer
